+++ lib/audio.c
@@ -123,7 +123,7 @@ ml_function_noarg(al_stop_samples)
 CAMLprim value ml_al_load_audio_stream(value filename, value buffer_count, value samples)
 {
     CAMLparam3(filename, buffer_count, samples);
-    ALLEGRO_AUDIO_STREAM * c_stream = al_load_audio_stream(
+    ALLEGRO_AUDIO_STREAM *c_stream = al_load_audio_stream(
         String_val(filename), Long_val(buffer_count), Int_val(samples));
     CAMLreturn(Val_ptr(c_stream));
 }