+++ CHANGES.md
@@ -1,7 +1,11 @@
 # 0.2
 
 - Audio: support for a basic API.
+- Displays: support for size and position.
 - Joystick: full support.
+- Shader: full support.
+- System: support for standard paths.
+- Primitives addon: support for spline and ribbon.
 
 
 # 0.1 (2025-02)